增加了大量功能，修复了大量bug

+++ include/DPIManager.h
@@ -69,6 +69,6 @@ private:
     
     // Helper functions
     float clampScale(float scale) const;
-    int clampFontSize(int size) const;
+    float clampFontSize(float size) const;
     float clampLineWidth(float width) const;
 };
